requirements.txt
# Core data science stack
numpy>=1.24.4
pandas>=2.0.3
scikit-learn>=1.3.0
joblib>=1.3.2

# NLP
spacy>=3.6.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
nltk>=3.8.1
sumy>=0.11.0

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system=="Windows"

# ML and visualization
scipy>=1.10.1
matplotlib>=3.7.2
seaborn>=0.12.2

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm==4.65.0




PS D:\ml> python legal_doc_analyzer_ml.py "D:\ml\dataset\dataset\IN-Ext\judgement\1953_L_1.txt"  
PS D:\ml> python legal_doc_analyzer_ml.py "D:\ml\dataset\dataset\IN-Ext\judgement\2012_S_270.txt"  
PS D:\ml> python test_summarizer.py                                                                     